fix: add yjs dependency for memory sync
