+++ package.json
@@ -75,6 +75,7 @@
     "three": "^0.179.1",
     "vaul": "1.1.2",
     "vite-plugin-pwa": "^0.20.5",
+    "yjs": "^13.6.6",
     "zod": "^3.25.76",
     "zustand": "^4.5.7"
   },
