+++ src/components/avatar/AuroraSphere.stories.tsx
@@ -1,0 +1,10 @@
+import React from 'react';
+import { AuroraSphere } from './AuroraSphere';
+
+export default {
+  title: 'Avatar/AuroraSphere',
+  component: AuroraSphere,
+};
+
+export const Default = () => <AuroraSphere />;
+export const Speaking = () => <AuroraSphere speaking />;
