--- src/components/avatar/EvolvingSphere.tsx
@@ -1,2 +1,0 @@
-// Re-export the existing evolving sphere from effects to match requested path
-export { EvolvingSphere } from "@/components/effects/EvolvingSphere";
