+++ src/index.css
@@ -465,87 +465,6 @@ All colors MUST be HSL.
 /* Safe area helper for HUD */
 .pb-safe { padding-bottom: var(--edge-bottom); }
 
-/* ===== EvolvingSphere (gooey liquid orb) ===== */
-
-.es-wrap {
-  width: var(--es-size);
-  height: var(--es-size);
-  position: relative;
-  filter: url(#es-goo);
-}
-
-.es-svg { position:absolute; width:0; height:0; }
-
-.es-sphere {
-  position: relative;
-  width: 100%; height: 100%;
-  border-radius: 50%;
-  /* subtle outer glow */
-  box-shadow:
-    0 0 24px hsla(var(--es-hue), var(--es-sat), calc(var(--es-light) + 10%), .35),
-    inset 0 0 18px hsla(var(--es-hue), var(--es-sat), calc(var(--es-light) + 20%), .45);
-  opacity: var(--es-intensity, 1);
-}
-
-/* core blob (breathes) */
-.es-core {
-  position:absolute; inset: 0;
-  border-radius: 50%;
-  background:
-    radial-gradient(60% 60% at 35% 35%, hsla(var(--es-hue), var(--es-sat), 92%, .9), transparent 55%),
-    radial-gradient(90% 90% at 50% 50%, hsla(var(--es-hue), var(--es-sat), var(--es-light), 1), hsla(var(--es-hue), var(--es-sat), calc(var(--es-light) - 10%), 1));
-  animation: es-breathe calc(var(--es-speed) / var(--es-intensity, 1)) var(--ease-standard) infinite;
-}
-@keyframes es-breathe {
-  0%   { transform: scale(1); }
-  50%  { transform: scale(1.06); }
-  100% { transform: scale(1); }
-}
-
-/* specular highlight (glassy top-left) */
-.es-spec {
-  position:absolute;
-  width: 38%; height: 38%;
-  left: 12%; top: 10%;
-  border-radius: 50%;
-  background: radial-gradient(60% 60% at 40% 40%, hsl(0 0% 100% / .9), hsl(0 0% 100% / .1) 60%, transparent 75%);
-  filter: blur(1px);
-  pointer-events:none;
-}
-
-/* satellites (metaballs orbit) */
-.es-sat {
-  --r: calc(var(--es-size) * 0.48);  /* orbit radius */
-  --angle: calc(360deg / 6 * var(--i)); /* distribute */
-  position:absolute; left:50%; top:50%;
-  width: calc(var(--es-size) * 0.26);
-  height: calc(var(--es-size) * 0.26);
-  margin: calc(var(--es-size) * -0.13) 0 0 calc(var(--es-size) * -0.13);
-  border-radius: 50%;
-  background:
-    radial-gradient(60% 60% at 35% 35%, hsla(var(--es-hue), var(--es-sat), 96%, .75), transparent 60%),
-    radial-gradient(90% 90% at 50% 50%, hsla(var(--es-hue), var(--es-sat), calc(var(--es-light) + 6%), 1), hsla(var(--es-hue), var(--es-sat), calc(var(--es-light) - 6%), 1));
-  animation:
-    es-orbit calc(var(--es-speed) / var(--es-intensity, 1)) linear infinite,
-    es-jelly calc((var(--es-speed) * .6) / var(--es-intensity, 1)) var(--ease-standard) infinite;
-  transform-origin: calc(var(--r)) 0;
-}
-@keyframes es-orbit {
-  0%   { transform: rotate(var(--angle)) translateX(var(--r)) rotate(0deg); }
-  100% { transform: rotate(calc(var(--angle) + 360deg)) translateX(var(--r)) rotate(-360deg); }
-}
-@keyframes es-jelly {
-  0%,100% { border-radius: 50% 46% 52% 48%/50% 52% 48% 50%; }
-  50%     { border-radius: 48% 54% 46% 52%/54% 46% 52% 48%; }
-}
-
-/* small ring when hovered (if you want in lists) */
-.es-wrap:hover .es-sphere {
-  box-shadow:
-    0 0 28px hsla(var(--es-hue), var(--es-sat), calc(var(--es-light) + 12%), .45),
-    inset 0 0 18px hsla(var(--es-hue), var(--es-sat), calc(var(--es-light) + 22%), .5);
-}
-
 /* Room theming via CSS variables (tints) */
 .room-control { --primary: 203 89% 53%; --accent: 203 89% 53%; }
 .room-focus { --primary: 222 87% 55%; --accent: 222 87% 55%; }
