+++ src/styles/hud-maple.css
@@ -9,7 +9,7 @@
   color: hsl(var(--foreground));
 }
 
-/* Optional placeholder orb (we use EvolvingSphere in code, but keep this for fallback) */
+/* Optional placeholder orb (AuroraSphere used in code; this is fallback) */
 .avatar-sphere { width: 42px; height: 42px; border-radius: 999px; background: radial-gradient(circle at 30% 30%, hsl(var(--accent) / 1) 0%, hsl(var(--primary) / 1) 60%, hsl(var(--primary) / 0.9) 100%); box-shadow: inset 0 10px 18px hsl(0 0% 100% / 0.25), 0 8px 20px hsl(var(--primary) / 0.45); }
 .avatar-spark { position:absolute; right:-2px; top:-2px; width:10px; height:10px; border-radius:999px; background: radial-gradient(circle, hsl(45 100% 70%) 0%, hsl(30 100% 55%) 70%); box-shadow: 0 0 12px hsl(40 100% 65%); }
